% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{add_plugin_eslint}
\alias{add_plugin_eslint}
\title{ESLint Plugin}
\usage{
add_plugin_eslint()
}
\description{
Add the \href{https://www.npmjs.com/package/eslint-webpack-plugin}{eslint-webpack-plugin}
run ESLint on files.
}
