% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{add_plugin_jsdoc}
\alias{add_plugin_jsdoc}
\title{Add Plugin jsdoc}
\usage{
add_plugin_jsdoc(edit = interactive())
}
\arguments{
\item{edit}{Whether to edit the configuration file.}
}
\description{
Add the \href{https://github.com/jsdoc/jsdoc}{jsdoc}
plugin to generate documentation from JavaScript code
with tags similar to roxygen2.
}
