% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{add_plugin_prettier}
\alias{add_plugin_prettier}
\title{Prettier Plugin}
\usage{
add_plugin_prettier()
}
\description{
Add the \href{https://www.npmjs.com/package/prettier-webpack-plugin}{prettier-webpack-plugin} to
prettify the pre-bundled files.
}
