% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework7.R
\name{apply_framework7}
\alias{apply_framework7}
\title{Apply Framework7}
\usage{
apply_framework7()
}
\description{
Apply Framework7 to a project, adds the relevant (babel) loader, installs dependencies,
and creates, or updates, or replaces the \code{srcjs/index.js} file.
}
\details{
After running this function and bundling the JavaScript remember to place
\verb{div(id = "app"), tags$script(src = "www/index.js")} at the bottom of your shiny UI.
}
