% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vue.R
\name{apply_vue}
\alias{apply_vue}
\title{Apply Vue}
\usage{
apply_vue(replace_index = FALSE)
}
\arguments{
\item{replace_index}{Whether to replace the \code{srcjs/index.js} file.}
}
\description{
Apply Vue to a project, adds the relevant (babel) loader, installs dependencies,
and creates, or updates, or replaces the \code{srcjs/index.js} file.
}
\details{
After running this function and bundling the JavaScript remember to place
\verb{div(id = "app"), tags$script(src = "www/index.js")} at the bottom of your shiny UI.
}
\examples{
\dontrun{
golem::create_golem("useVue")
packer::scaffold_golem(vue = TRUE)
}

}
