% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\alias{build_widget}
\alias{watch_widget}
\title{Build & Develop}
\usage{
build_widget(mode = c("production", "development", "none"))

watch_widget()
}
\arguments{
\item{mode}{The configuration mode tells webpack to use
its built-in optimizations accordingly.}
}
\description{
Build and develop the widget.
}
\section{Functions}{

\itemize{
\item \code{\link[=build_widget]{build_widget()}} - builds the widget using webpack to produce the required JavaScript file in the \code{inst/htmlwidgets} directory.
\item \code{\link[=watch_widget]{watch_widget()}} - watches for changes in the \code{srcjs} and rebuilds if necessary.
}
}

\examples{
\dontrun{watch_widget()}

}
