% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checks}
\alias{checks}
\title{Checks}
\usage{
checks()
}
\description{
Run checks on a package using packer.
}
\section{Checks}{

\itemize{
\item Output files are minified
\item \link{put_precommit_hook} is in place
\item \link{put_rprofile_adapt} is in place
}
}

