% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{engine}
\alias{engine}
\alias{engine_set}
\alias{engine_get}
\title{Set Engine}
\usage{
engine_set(engine = c("npm", "yarn"))

engine_get()
}
\arguments{
\item{engine}{The engine to use, npm or yarn.}
}
\description{
Defines the engine to use with packer.
One can pick between npm and yarn.
}
\details{
Generally one would want to define
the engine prior to scaffolding.
For convenience you can instead set the environment
variable \code{PACKER_ENGINE} to your engine of choice.
Packer reads this variable, all subsequent use
of packer will use the defined engine.
You can use the function \code{usethis::edit_r_environ}
to do so.
}
