% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{include_action_check}
\alias{include_action_check}
\title{Github Actions}
\usage{
include_action_check()
}
\description{
Adds a Github Action to the package that will bundle
the JavaScript for production on commit.
}
