% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_minified}
\alias{is_minified}
\title{Is minified}
\usage{
is_minified(dir)
}
\arguments{
\item{dir}{Directory of files to check}
}
\description{
Checks whether files within a directory are minified.
}
\examples{
\dontrun{is_minified("path/to/js-dir")}

}
