% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{jsdoc}
\alias{jsdoc}
\alias{add_plugin_jsdoc}
\alias{add_jsdoc_tutorial}
\title{Add Plugin jsdoc}
\usage{
add_plugin_jsdoc(edit = interactive())

add_jsdoc_tutorial(name, edit = interactive())
}
\arguments{
\item{edit}{Whether to open relevent file.}

\item{name}{Name of tutorial}
}
\description{
Add the \href{https://github.com/jsdoc/jsdoc}{jsdoc}
plugin to generate documentation from JavaScript code
with tags similar to roxygen2.
}
