% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{make_library}
\alias{make_library}
\title{Make Library}
\usage{
make_library(name, type = "umd")
}
\arguments{
\item{name}{Name of the library, see details.}

\item{type}{Type of the library.}
}
\description{
Adds library settings to webpack config.
This allow exporting JavaScript objects.
}
\details{
The functions will be exported
at the specified \code{name}, e.g.: if
the name is \code{myLib} then functions can be
called with \code{myLib.function();}.
}
