% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mockup}
\alias{mockup}
\alias{tmp_package}
\alias{tmp_golem}
\alias{tmp_project}
\alias{tmp_ambiorix}
\alias{tmp_delete}
\title{Mock up}
\usage{
tmp_package()

tmp_golem()

tmp_project()

tmp_ambiorix()

tmp_delete(tmp)
}
\arguments{
\item{tmp}{A temp mock up project.}
}
\description{
Functions to mock up packages for tests
}
