% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{npm_fix}
\alias{npm_fix}
\title{Audit Fix}
\usage{
npm_fix()
}
\description{
Scan your project for vulnerabilities and automatically install
any compatible updates to vulnerable dependencies.
}
\details{
Runs \verb{npm audit fix}
}
