% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{npm_install}
\alias{npm_install}
\title{Install}
\usage{
npm_install(..., scope = c("local", "dev", "global"))
}
\arguments{
\item{...}{Packages to install.}

\item{scope}{Scope of installation, see scopes.}
}
\description{
Install npm packges.
}
\section{Scopes}{

\itemize{
\item \code{local} - Installs packages for project with \code{--save}
\item \code{dev} - Installs dev packages for project with \code{--save-dev}
\item \code{global} - Instals packages globally with \code{-g}
}
}

\examples{
# install browserify globally
\dontrun{npm_install("browserify", scope = "global")}

}
