% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packer-package.R
\docType{package}
\name{packer-package}
\alias{packer}
\alias{packer-package}
\title{packer: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahasverus/packer}
  \item Report bugs at \url{https://github.com/ahasverus/packer/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{rdev.nc@gmail.com} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
