% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{put_precommit_hook}
\alias{put_precommit_hook}
\title{Put Pre-Commit Hook}
\usage{
put_precommit_hook()
}
\description{
Add a pre-commit hook that runs at every commit
to ensure that JavaScript files are minified.
}
\note{
Will only work if using git.
}
\examples{
\dontrun{put_precommit_hook()}

}
