% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{put_recommended}
\alias{put_recommended}
\title{Recommended Checks}
\usage{
put_recommended()
}
\description{
Recommended checks for packer projects,
runs \link{put_rprofile_adapt} and
\link{put_precommit_hook}.
}
