% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{put_rprofile_adapt}
\alias{put_rprofile_adapt}
\title{Rprofile}
\usage{
put_rprofile_adapt()
}
\description{
Add \code{\link[=engine_adapt]{engine_adapt()}} to \code{.Rprofile}.
}
\details{
This is recommended so anyone contributing to
the project is guaranteed to be on the correct
engine.
}
