% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{scaffold_extension}
\alias{scaffold_extension}
\title{Shiny Extension}
\usage{
scaffold_extension(name)
}
\arguments{
\item{name}{Name of extension used to define file names and functions.}
}
\value{
\code{TRUE} (invisibly) if successfully run.
}
\description{
Creates the basic structure for a shiny extension.
}
