% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem.R
\name{scaffold_golem}
\alias{scaffold_golem}
\title{Golem}
\usage{
scaffold_golem(react = FALSE, vue = FALSE, edit = interactive())
}
\arguments{
\item{react}{Whether to include React, internally runs \code{\link[=apply_react]{apply_react()}} and adapts the \code{srcjs/index.js} template for React.}

\item{vue}{Whether to include Vue, internally runs \code{\link[=apply_vue]{apply_vue()}} and adapts the \code{srcjs/index.js} template for Vue.}

\item{edit}{Automatically open pertinent files.}
}
\value{
\code{TRUE} (invisibly) if successfully run.
}
\description{
Creates the basic structure for golem app with JavaScript.
}
\details{
Only one of \code{react} or \code{vue} can be set to \code{TRUE}.
}
