% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{scaffold_input}
\alias{scaffold_input}
\title{Scaffold a Custom Input}
\usage{
scaffold_input(name)
}
\arguments{
\item{name}{Name of input, will define internal name binding and CSS class.}
}
\description{
Sets basic structure for a shiny input.
}
\examples{
\dontrun{scaffold_input("increment")}

}
