% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leprechaun.R
\name{scaffold_leprechaun}
\alias{scaffold_leprechaun}
\title{Leprechaun}
\usage{
scaffold_leprechaun(
  react = FALSE,
  vue = FALSE,
  use_cdn = TRUE,
  edit = interactive()
)
}
\arguments{
\item{react}{Whether to include React, internally runs \code{\link[=apply_react]{apply_react()}}
and adapts the \code{srcjs/index.js} template for React.}

\item{vue}{Whether to include Vue, internally runs \code{\link[=apply_vue]{apply_vue()}} and
adapts the \code{srcjs/index.js} template for Vue.}

\item{use_cdn}{Whether to use the CDN for react or vue dependencies,
this is passed to \code{\link[=apply_react]{apply_react()}} or \code{\link[=apply_vue]{apply_vue()}} if \code{react} or
\code{vue} arguments are set to \code{TRUE} and ignored otherwise.}

\item{edit}{Automatically open pertinent files.}
}
\value{
\code{TRUE} (invisibly) if successfully run.
}
\description{
Creates the basic structure for leprechaun app with JavaScript.
}
\details{
Only one of \code{react} or \code{vue} can be set to \code{TRUE}.
}
