% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{set_npm}
\alias{set_npm}
\title{Set npm path}
\usage{
set_npm(path = NULL)
}
\arguments{
\item{path}{Path to npm installation to use.}
}
\description{
By default packer looks for the npm installation using
the \code{which} (or \code{where}) command.
This function lets you override that behaviour and
force a specific npm installation.
}
