% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yarn.R
\name{set_yarn}
\alias{set_yarn}
\title{Set yarn path}
\usage{
set_yarn(path = NULL)
}
\arguments{
\item{path}{Path to yarn installation to use.}
}
\description{
By default packer looks for the yarn installation using
the \code{which} (or \code{where}) command.
This function lets you override that behaviour and
force a specific yarn installation.
}
