% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{style_loaders}
\alias{style_loaders}
\alias{use_css}
\alias{use_sass}
\title{Use Styles}
\usage{
use_css()

use_sass()
}
\description{
Installs loaders and adds relevant configuration rules to \code{srcjs/config/loaders.json}.
}
\details{
This will let you import styles much like any other modules, e.g.: \verb{import './styles.css'}.
}
\section{Packages}{

\itemize{
\item \code{\link[=use_css]{use_css()}} - installs \code{style-loader} and \code{css-loader} packages as dev.
\item \code{\link[=use_sass]{use_sass()}} - installs \code{style-loader}, \code{css-loader}, \code{sass-loader}, and \code{sass} as dev.
}
}

