% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typescript.R
\name{types}
\alias{types}
\alias{ts_get_types}
\alias{ts_get_type}
\title{Install Types}
\usage{
ts_get_types(..., versions = NULL)

ts_get_type(type, version = NULL)
}
\arguments{
\item{...}{Types to install.}

\item{versions, version}{Corresponding versions of types passed
to \code{...}, if \code{NULL} the latest version is installed.}

\item{type}{Name of types \verb{@types/*} to install.}
}
\description{
Install TypeScript types from npm.
}
\section{Functions}{

\itemize{
\item \code{ts_get_types}: Flexible to retrieve multiple types.
\item \code{ts_get_type}: Convenience to easily retrieve a single type.
}
}

\examples{
\dontrun{ts_get_type("jquery")}
\dontrun{ts_get_types("@types/jquery")}

}
