% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_css}
\alias{use_css}
\title{Use Styles}
\usage{
use_css()
}
\description{
Installs loaders, creates \code{srcjs/styles} directory containing a CSS file and prints required modifications to the \code{webpack.config.js} file.
}
\details{
The modifications to the \code{webpack.config.js} must be placed in the JSON under \code{module.exports}.
The created CSS file can then be imported much like any other JavaScript file.
}
