% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_loader_babel}
\alias{use_loader_babel}
\title{Use babel Loader}
\usage{
use_loader_babel(test = "\\\\.(js|jsx)$")
}
\arguments{
\item{test}{Test regular expression test which files should be transformed by the loader.}
}
\description{
Adds the loader for babel comiler to the loader configuration file.
}
\details{
Excludes \code{node_modules} by default.
}
