% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_loader_coffee}
\alias{use_loader_coffee}
\title{Use Coffee Loader}
\usage{
use_loader_coffee(test = "\\\\.coffee$")
}
\arguments{
\item{test}{Test regular expression test which files should be transformed by the loader.}
}
\description{
Adds the \href{https://webpack.js.org/loaders/coffee-loader/}{\code{coffee-loader}} to use
cofeescript.
}
\details{
Excludes \code{node_modules} by default.
}
