% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_loader_eslint}
\alias{use_loader_eslint}
\title{Use ESlint}
\usage{
use_loader_eslint(test = "\\\\.(js|jsx)$")
}
\arguments{
\item{test}{Test regular expression test which files should be transformed by the loader.}
}
\description{
Adds the \href{https://github.com/webpack-contrib/eslint-loader}{\code{eslint-loader}}
to resolve files: \code{png}, \code{jpg}, \code{jpeg}, and \code{gif}.
}
