% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_loader_file}
\alias{use_loader_file}
\title{Use File Loader}
\usage{
use_loader_file(test = "\\\\.(png|jpe?g|gif)$/i")
}
\arguments{
\item{test}{Test regular expression test which files should be transformed by the loader.}
}
\description{
Adds the \href{https://webpack.js.org/loaders/file-loader/}{\code{file-loader}}
to resolve files: \code{png}, \code{jpg}, \code{jpeg}, and \code{gif}.
}
