% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_loader_framework7}
\alias{use_loader_framework7}
\title{Use Framework7 Loader}
\usage{
use_loader_framework7(test = "\\\\.(f7).(html|js|jsx)$")
}
\arguments{
\item{test}{Test regular expression test which files should be transformed by the loader.}
}
\description{
Adds the \href{https://www.npmjs.com/package/framework7-loader}{Framework7 loader}.
}
\details{
Excludes \code{node_modules} by default. If used outside \code{scaffold_golem}
context, installs the babel-loader in the dev scope.
}
