% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_loader_mocha}
\alias{use_loader_mocha}
\title{Use Mocha Loader}
\usage{
use_loader_mocha(test = "\\\\.test\\\\.js$")
}
\arguments{
\item{test}{Test regular expression test which files should be transformed by the loader.}
}
\description{
Adds the \href{https://webpack.js.org/loaders/mocha-loader/}{\code{mocha-loader}} for tests.
}
\details{
Excludes \code{node_modules} by default.
}
