% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_loader_pug}
\alias{use_loader_pug}
\title{Use Pug Loader}
\usage{
use_loader_pug(test = "\\\\.pug$")
}
\arguments{
\item{test}{Test regular expression test which files should be transformed by the loader.}
}
\description{
Adds the loader for the pug templating engine.
}
