% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_loader_rule}
\alias{use_loader_rule}
\title{Add a Loader Ruée}
\usage{
use_loader_rule(
  packages,
  test,
  ...,
  use = as.list(packages),
  .name_use = "use"
)
}
\arguments{
\item{packages}{NPM packages (loaders) to install.}

\item{test}{Test regular expression test which files should be transformed by the loader.}

\item{...}{Any other options to pass to the rule.}

\item{use}{Name of the loaders to use for \code{test}.}

\item{.name_use}{Dependending on the webpack config
one might want to change the \code{use} to \code{loader} or \code{loaders}.}
}
\description{
Adds a loader rule that is not yet implemened in packer.
}
\details{
Reads the \code{srcsjs/config/loaders.json} and appends the rule.
}
