% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_loader_svelte}
\alias{use_loader_svelte}
\title{Use Svelte Loader}
\usage{
use_loader_svelte(test = "\\\\.(html|svelte)$")
}
\arguments{
\item{test}{Test regular expression test which files should be transformed by the loader.}
}
\description{
Add the loader svelte
}
