% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_loader_ts}
\alias{use_loader_ts}
\title{Use Typescript Loader}
\usage{
use_loader_ts(test = "\\\\.tsx?$")
}
\arguments{
\item{test}{Test regular expression test which files should be transformed by the loader.}
}
\description{
Adds the loader for the pug templating engine.
}
