% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_loader_vue}
\alias{use_loader_vue}
\title{Use Vue Loader}
\usage{
use_loader_vue(test = "\\\\.vue$")
}
\arguments{
\item{test}{Test regular expression test which files should be transformed by the loader.}
}
\description{
Adds the Vue loader to the loader configuration file.
}
\details{
Every time a new version of Vue is released, a corresponding version of \code{vue-template-compiler}
is released together. The compiler's version must be in sync with the base Vue package so that \code{vue-loader}
produces code that is compatible with the runtime. This means every time you upgrade Vue in your project,
you should upgrade \code{vue-template-compiler} to match it as well.
}
