% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_loader_vue_style}
\alias{use_loader_vue_style}
\title{Use Vue Style Loader}
\usage{
use_loader_vue_style()
}
\description{
Similar to \code{\link[=use_loader_css]{use_loader_css()}}, dynamically inject CSS into the document as style tags.
}
