% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{use_npm}
\alias{use_npm}
\title{Use npm}
\usage{
use_npm(path = NULL)
}
\arguments{
\item{path}{Path to npm installation to use.}
}
\description{
By default packer looks for the npm installation using the \code{which} command.
This function lets you override that behaviour and force a specific npm installation.
}
\examples{
\dontrun{use_npm("/usr/local/bin/npm")}

}
