% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind.R
\name{use_tailwind}
\alias{use_tailwind}
\title{Use Tailwind}
\usage{
use_tailwind(test = "\\\\.css$")
}
\arguments{
\item{test}{Test regular expression test which files should be transformed by the loader.}
}
\description{
Creates PostCSS, and tailwindcss config files as well
as adds the appropriate loaders and installs dependencies.
}
