% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yarn.R
\name{yarn_cache}
\alias{yarn_cache}
\title{Yarn cache}
\usage{
yarn_cache(...)
}
\arguments{
\item{...}{Arguments to pass to \verb{yarn cache}.}
}
\value{
The semver as a string.
}
\description{
Get the version of yarn.
}
\examples{
\dontrun{yarn_cache("clean")}

}
