% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yarn.R
\name{yarn_global}
\alias{yarn_global}
\alias{engine_yarn_install}
\alias{engine_yarn_set}
\title{Yarn Global}
\usage{
engine_yarn_install()

engine_yarn_set(version = "latest")
}
\arguments{
\item{version}{Version to set yarn}
}
\description{
Installs or manage yarn \emph{globally}.
}
\section{Functions}{

\itemize{
\item \code{engine_yarn_install}: Installs yarn globally.
\item \code{engine_yarn_set}: Set yarn function.
}
}

\examples{
\dontrun{engine_yarn_install()}

}
