% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yarn.R
\name{yarn_install}
\alias{yarn_install}
\alias{yarn_add}
\alias{yarn_remove}
\title{Install and Uninstall yarn Packages}
\usage{
yarn_add(..., scope = c("dev", "prod"))

yarn_install()

yarn_remove(..., scope = c("dev", "prod"))
}
\arguments{
\item{...}{Packages to install or uninstall.}

\item{scope}{Scope of installation or uninstallation, see scopes.}
}
\description{
Install and uninstall yarn packages.
}
\section{Scopes}{

\itemize{
\item \code{prod} - Add/remove packages for project with no flag
\item \code{dev} - Installs/Uninstalls dev packages for project with \code{--dev}
}
}

\examples{
\dontrun{yarn_add("browserify")}

}
