% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R, R/yarn.R
\name{yarn_version}
\alias{yarn_version}
\title{Npm version}
\usage{
yarn_version()

yarn_version()
}
\value{
The semver as a string.

The semver as a string.
}
\description{
Get the version of npm.

Get the version of yarn.
}
\examples{
\dontrun{yarn_version()}

}
