% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{add_plugin_html}
\alias{add_plugin_html}
\title{HTML Plugin}
\usage{
add_plugin_html(use_pug = FALSE)
}
\arguments{
\item{use_pug}{Set to \code{TRUE} to use the \href{https://pugjs.org/}{pug engine}.}
}
\description{
Add the \href{https://webpack.js.org/plugins/html-webpack-plugin/}{htmlWebpackPlugin} to
the configuration to generate HTML with webpack.
}
