% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vue.R
\name{apply_vue}
\alias{apply_vue}
\title{Apply Vue}
\usage{
apply_vue(use_cdn = TRUE)
}
\arguments{
\item{use_cdn}{Whether to use the CDN for \code{vue} (recommended).
This means later importing the dependencies in the shiny UI using \code{vueCDN()},
this function will be created in a \code{R/vue_cdn.R}.
The correct instructions are printed to the console by the application.}
}
\description{
Apply Vue to a project, adds the relevant (babel) loader, installs dependencies,
and creates, or updates, or replaces the \code{srcjs/index.js} file.
}
\details{
After running this function and bundling the JavaScript remember to place
\verb{div(id = "app"), tags$script(src = "www/index.js")} at the bottom of your shiny UI.
}
