% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\alias{build_packer}
\alias{watch_packer}
\title{Build & Watch}
\usage{
build_packer(mode = c("production", "development", "none"))

watch_packer()
}
\arguments{
\item{mode}{The configuration mode tells webpack to use
its built-in optimisations accordingly.}
}
\description{
Build and watch the JavaScript.
}
\section{Functions}{

\itemize{
\item \code{\link[=build_packer]{build_packer()}} - builds the JavaScript using webpack to produce the bundle in the appropriate directory.
\item \code{\link[=watch_packer]{watch_packer()}} - watches for changes in the \code{srcjs} and rebuilds if necessary.
}
}

\examples{
\dontrun{watch_widget()}

}
