% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{build}
\alias{build}
\alias{bundle}
\alias{watch}
\title{bundle & Watch}
\usage{
bundle(mode = c("production", "development", "none"))

watch()
}
\arguments{
\item{mode}{The configuration mode tells webpack to use
its built-in optimisations accordingly.}
}
\description{
Bundle and watch the JavaScript.
}
\section{Functions}{

\itemize{
\item \code{\link[=bundle]{bundle()}} - bundle the JavaScript using webpack to produce the bundle in the appropriate directory.
\item \code{\link[=watch]{watch()}} - watches for changes in the \code{srcjs} and rebuilds if necessary.
}
}

\examples{
\dontrun{watch_widget()}

}
