% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{engine}
\alias{engine}
\alias{engine_set}
\alias{engine_get}
\alias{engine_adapt}
\alias{engine_which}
\title{Set Engine}
\usage{
engine_set(engine = c("npm", "yarn"))

engine_get()

engine_adapt()

engine_which()
}
\arguments{
\item{engine}{The engine to use, npm or yarn.}
}
\description{
Defines the engine to use with packer.
One can pick between npm and yarn.
}
\details{
Generally one would want to define
the engine prior to scaffolding.
For convenience you can instead set the environment
variable \code{PACKER_ENGINE} to your engine of choice.
Packer reads this variable, all subsequent use
of packer will use the defined engine.
You can use the function \code{usethis::edit_r_environ}
to do so.
}
\section{Functions}{

\itemize{
\item \code{engine_set}: Define the engine to use for the project.
\item \code{engine_get}: Retrieve the default engine.
\item \code{engine_which}: Retrieve which engine the project is set
to use-.
\item \code{engine_adapt}: Change the engine to match that of the
poject.
}
}

