% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{make_library}
\alias{make_library}
\title{Make Library}
\usage{
make_library(name = "[name]", type = "umd")
}
\arguments{
\item{name}{Name of the library, default recommended,
see details.}

\item{type}{Type of the library.}
}
\description{
Adds library settings to webpack config.
This allow exporting JavaScript objects.
}
\details{
The functions will be exported
at the specified \code{name}, e.g.: if
the name is \code{myLib} then functions can be
called with \code{myLib.function();}.
The default (\verb{[name]}) means the name of the exported
library will be the same as the name of the scaffold.
This is advised because otherwise, if one has multiple
scaffold, an absolute will overwrite itself and only
the last scaffold added will be a valid library.
}
