% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{npm_install}
\alias{npm_install}
\alias{npm_uninstall}
\title{Install and Uninstall Npm Packages}
\usage{
npm_install(..., scope = c("dev", "prod", "global"))

npm_uninstall(..., scope = c("dev", "prod", "global"))
}
\arguments{
\item{...}{Packages to install or uninstall.}

\item{scope}{Scope of installation or uninstallation, see scopes.}
}
\description{
Install and uninstall npm packages.
}
\section{Scopes}{

\itemize{
\item \code{prod} - Installs/Uninstalls packages for project with \code{--save}
\item \code{dev} - Installs/Uninstalls dev packages for project with \code{--save-dev}
\item \code{global} - Installs/Uninstalls packages globally with \code{-g}
}
}

