% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bare.R
\name{scaffold_bare}
\alias{scaffold_bare}
\title{Bare}
\usage{
scaffold_bare(edit = interactive())
}
\arguments{
\item{edit}{Automatically open pertinent files.}
}
\value{
\code{TRUE} (invisibly) if successfully run.
}
\description{
Creates a bare scaffold for not specific use case, as
opposed to other scaffolds. This scaffold does not
generate R code.
}
\examples{
if(interactive()){
# current directory
wd <- getwd()

# create a mock up ambiorix project
tmp <- tmp_package()

# move to package
setwd(tmp)

# scaffold bare
scaffold_bare()

# clean up
setwd(wd)
tmp_delete(tmp)
}

}
