% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{scaffold_extension}
\alias{scaffold_extension}
\title{Shiny Extension}
\usage{
scaffold_extension()
}
\value{
\code{TRUE} (invisibly) if successfully run.
}
\description{
Creates the basic structure for a shiny extension.
}
