% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{scaffold_extension}
\alias{scaffold_extension}
\title{Shiny Extension}
\usage{
scaffold_extension(name, edit = interactive())
}
\arguments{
\item{name}{Name of extension used to define file names and functions.}

\item{edit}{Automatically open pertinent files.}
}
\value{
\code{TRUE} (invisibly) if successfully run.
}
\description{
Creates the basic structure for a shiny extension.
}
\examples{
if(interactive()){
# current directory
wd <- getwd()

# create a mock up ambiorix project
tmp <- tmp_package()

# move to package
setwd(tmp)

# scaffold ambiorix
scaffold_extension()

# clean up
setwd(wd)
tmp_delete(tmp)
}

}
