% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem.R
\name{scaffold_golem}
\alias{scaffold_golem}
\title{Golem}
\usage{
scaffold_golem(react = FALSE, edit = interactive())
}
\arguments{
\item{react}{Whether to include react, internally runs \code{\link[=apply_react]{apply_react()}} and adapts the \code{srcjs/index.js} template for React.}

\item{edit}{Automatically open pertinent files.}
}
\value{
\code{TRUE} (invisibly) if successfully run.
}
\description{
Creates the basic structure for golem app with JavaScript.
}
