% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem.R
\name{scaffold_golem}
\alias{scaffold_golem}
\title{Golem}
\usage{
scaffold_golem(edit = interactive(), verbose = FALSE)
}
\arguments{
\item{edit}{Automatically open the widget's JavaScript source files after
creating the scaffold.}

\item{verbose}{Whether to print every output (e.g.: npm console output),
if \code{FALSE} only print minimalistic messages.}
}
\value{
\code{TRUE} (invisibly) if successfully run.
}
\description{
Creates the basic structure for golem app with JavaScript.
}
