% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{scaffold_input}
\alias{scaffold_input}
\title{Scaffold a Custom Input}
\usage{
scaffold_input(name, edit = interactive())
}
\arguments{
\item{name}{Name of input, will define internal name binding and CSS class.}

\item{edit}{Automatically open pertinent files.}
}
\value{
\code{TRUE} (invisibly) if successfully run.
}
\description{
Sets basic structure for a shiny input.
}
\examples{
if(interactive()){
# current directory
wd <- getwd()

# create a mock up ambiorix project
tmp <- tmp_package()

# move to package
setwd(tmp)

# scaffold ambiorix
scaffold_input()

# clean up
setwd(wd)
tmp_delete(tmp)
}

}
