% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{scaffold_output}
\alias{scaffold_output}
\title{Scaffold Shiny Output}
\usage{
scaffold_output(name, edit = NULL)
}
\arguments{
\item{name}{Name of output, will define internal name binding and CSS class.}

\item{edit}{Automatically open pertinent files. Defaults to \code{NULL}, which looks
for the environment variable \code{PACKER_EDIT} and opens the files specified there.
Otherwise takes a boolean.}
}
\value{
\code{TRUE} (invisibly) if successfully run.
}
\description{
Sets basic structure for a shiny input.
}
\examples{
if (interactive()) {
  # current directory
  wd <- getwd()

  # create a mock up ambiorix project
  tmp <- tmp_package()

  # move to package
  setwd(tmp)

  # scaffold ambiorix
  scaffold_output()

  # clean up
  setwd(wd)
  tmp_delete(tmp)
}

}
