% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{scaffold_output}
\alias{scaffold_output}
\title{Scaffold Shiny Output}
\usage{
scaffold_output(name)
}
\arguments{
\item{name}{Name of output, will define internal name binding and CSS class.}
}
\description{
Sets basic structure for a shiny input.
}
\examples{
\dontrun{scaffold_output("display")}

}
