% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{scaffold_widget}
\alias{scaffold_widget}
\title{Scaffold Widget}
\usage{
scaffold_widget(name, edit = interactive(), verbose = FALSE)
}
\arguments{
\item{name}{Name of widget, also passed to \code{\link[htmlwidgets:scaffoldWidget]{htmlwidgets::scaffoldWidget()}}.}

\item{edit}{Automatically open the widget's JavaScript source files after
creating the scaffold.}

\item{verbose}{Whether to print every output (e.g.: npm console output),
if \code{FALSE} only print minimalistic messages.}
}
\value{
\code{TRUE} (invisibly) if successfully run.
}
\description{
Creates basic structure for a widget.
}
\details{
Internally runs \code{\link[htmlwidgets:scaffoldWidget]{htmlwidgets::scaffoldWidget()}}.
}
\examples{
\dontrun{scaffold_widget()}

}
