% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{scaffold_widget}
\alias{scaffold_widget}
\title{Create Package}
\usage{
scaffold_widget(name)
}
\arguments{
\item{name}{Name of widget, also passed to \code{\link[htmlwidgets:scaffoldWidget]{htmlwidgets::scaffoldWidget()}}.}
}
\description{
Create a package with
}
\examples{
\dontrun{scaffold_widget()}

}
