% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{scaffold_widget}
\alias{scaffold_widget}
\title{Scaffold Widget}
\usage{
scaffold_widget(name, edit = interactive())
}
\arguments{
\item{name}{Name of widget, also passed to \code{\link[htmlwidgets:scaffoldWidget]{htmlwidgets::scaffoldWidget()}}.}

\item{edit}{Automatically open pertinent files.}
}
\value{
\code{TRUE} (invisibly) if successfully run.
}
\description{
Creates basic structure for a widget.
}
\details{
Internally runs \code{\link[htmlwidgets:scaffoldWidget]{htmlwidgets::scaffoldWidget()}} do not run it prior to this function.
}
\examples{
if(interactive()){
# current directory
wd <- getwd()

# create a mock up ambiorix project
tmp <- tmp_package()

# move to package
setwd(tmp)

# scaffold ambiorix
scaffold_widget()

# clean up
setwd(wd)
tmp_delete(tmp)
}

}
