% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{style_loaders}
\alias{style_loaders}
\alias{use_loader_css}
\alias{use_loader_sass}
\title{Use Styles}
\usage{
use_loader_css(test = "\\\\.css$")

use_loader_sass(test = "\\\\.s[ac]ss$/i")
}
\arguments{
\item{test}{Test regular expression test which files should be transformed by the loader.}
}
\description{
Installs loaders and adds relevant configuration rules to \code{srcjs/config/loaders.json}.
}
\details{
This will let you import styles much like any other modules, e.g.: \verb{import './styles.css'}.
}
\section{Packages}{

\itemize{
\item \code{\link[=use_loader_css]{use_loader_css()}} - installs and imports \code{style-loader} and \code{css-loader} packages as dev.
\item \code{\link[=use_loader_sass]{use_loader_sass()}} - installs and imports \code{style-loader}, \code{css-loader}, and \code{sass-loader} as dev.
}
}

