% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{use_css}
\alias{use_css}
\title{Use Styles}
\usage{
use_css()
}
\description{
Installs loaders and prints required modifications to the \code{webpack.config.js} file.
}
\details{
The modifications to the \code{webpack.config.js} must be placed in the JSON under \code{module.exports}.
You can then place, for instance, create \code{srcjs/style.css} and import it in \code{scrjs/index.js} with \verb{import './style.css';}.
}
