% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_css}
\alias{use_css}
\title{Use Styles}
\usage{
use_css()
}
\description{
Installs loaders, creates \code{srcjs/styles} directory containing a CSS file and adds module rule to the configuration.
}
\details{
This will let you import styles much like any other modules.
}
