% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{use_loader_babel}
\alias{use_loader_babel}
\title{Use babel Loader}
\usage{
use_loader_babel(test = "\\\\.(js|jsx)$", use_eslint = FALSE)
}
\arguments{
\item{test}{Test regular expression test which files should be transformed by the loader.}

\item{use_eslint}{Whether to also add the ESlint loader.}
}
\description{
Adds the loader for babel compiler to the loader configuration file.
}
\details{
The \code{use_elsint} argument is useful here as loaders have
to be defined in the correct order or files might be checked after
being processed by babel.

Excludes \code{node_modules} by default.
}
